Transport deterministic signal tracks in semantic IR
